+++ packages/ethereum-provider/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  ...require("@rpch/configs-eslint"),
+};
