+++ packages/ethereum-provider/README.md
@@ -1,0 +1,3 @@
+# RPCh ethereum provider adaptor
+
+## Description
