+++ packages/ethereum-provider/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@rpch/configs-typescript/nodejs-library.json",
+  "include": ["src"],
+  "exclude": ["node_modules"]
+}
