+++ packages/ethereum-provider/tsup.config.js
@@ -1,0 +1,1 @@
+module.exports = require("@rpch/configs-tsup");
