+++ packages/ethers/src/index.spec.ts
@@ -2,7 +2,7 @@ import assert from "assert";
 import * as fixtures from "@rpch/common/build/fixtures";
 import * as RPChCrypto from "@rpch/crypto-for-nodejs";
 import mockSdk from "@rpch/sdk/build/index.mock";
-import { RPChEthereumProvider, RPChProvider } from ".";
+import { RPChProvider } from ".";
 
 // mock HOPRd interactions
 jest.mock("@rpch/common", () => ({
@@ -53,51 +53,6 @@ describe("test index.ts", function () {
     it("should get block number", async function () {
       const blockNumber = await provider.getBlockNumber();
       assert.equal(blockNumber, 25135304);
-    });
-  });
-  describe("test rpch ethereum provider", function () {
-    const sdkStore = fixtures.createAsyncKeyValStore();
-    const provider = new RPChEthereumProvider(
-      PROVIDER_URL,
-      {
-        crypto: RPChCrypto,
-        client: "",
-        timeout: TIMEOUT,
-        discoveryPlatformApiEndpoint: DISCOVERY_PLATFORM_API_ENDPOINT,
-      },
-      sdkStore.set,
-      sdkStore.get
-    );
-
-    beforeAll(async function () {
-      await provider.sdk.start();
-    });
-
-    afterAll(async function () {
-      await provider.sdk.stop();
-    });
-
-    // hook to emulate responses from the exit node
-    provider.sdk = mockSdk(provider.sdk);
-
-    it("should get chain id", function (done) {
-      provider.send(
-        { id: 1, jsonrpc: "2.0", method: "eth_chainId", params: [] },
-        (err, res) => {
-          assert.equal(res?.result, 1);
-          done();
-        }
-      );
-    });
-
-    it("should get block number", function (done) {
-      provider.send(
-        { id: 1, jsonrpc: "2.0", method: "eth_blockNumber", params: [] },
-        (err, res) => {
-          assert.equal(res?.result, 25135304);
-          done();
-        }
-      );
     });
   });
 });
